requirements.txt
Flask==0.10.1
pyserial==3.4
flask-WTF==0.14.2
tinydb==3.13.0
flask_table==0.5.0
apscheduler==3.6.0
